update dev env for webpack
