+++ README.md
@@ -3,3 +3,15 @@
 Relativity Simulator Source Code.
 Used Three.js for rendering tool.
 WEBSITE: https://relativity-simulator.netlify.app/
+
+# How to contribute
+
+1. install [nodejs](https://nodejs.org)
+
+2. clone this repository
+
+3. run 
+  > npm install
+
+  > npm run dev
+
