--- main.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8">
-		<title>Relativity Simulator</title>
-		<link rel="stylesheet" href="main.css">
-    <script>
-      window.onbeforeunload = () => "만들어진 월드가 저장되지 않을 수 있습니다."
-    </script>
-    <!-- Global site tag (gtag.js) - Google Analytics -->
-    <script async src="https://www.googletagmanager.com/gtag/js?id=G-PF03MR37HK"></script>
-    <script>
-        window.dataLayer = window.dataLayer || [];
-        function gtag(){dataLayer.push(arguments);}
-        gtag('js', new Date());
-
-        gtag('config', 'G-PF03MR37HK');
-    </script>
-	</head>
-	<body>
-    <canvas id="c"></canvas>
-    <div id="clocklabels"></div>
-    <div id="instructions">
-      <p>Click to Play</p>
-    </div>
-    <div id="crossbar">
-      <p>+</p>
-    </div>
-    <div id="ui">
-      <div class="tiles">
-        <input type="radio" class="add" name="voxel" id="voxel1" value="1" checked><label for="voxel1" style="background-position: -0% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel2" value="2"><label for="voxel2" style="background-position: -100% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel3" value="3"><label for="voxel3" style="background-position: -200% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel4" value="4"><label for="voxel4" style="background-position: -300% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel5" value="5"><label for="voxel5" style="background-position: -400% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel6" value="6"><label for="voxel6" style="background-position: -500% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel7" value="7"><label for="voxel7" style="background-position: -600% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel8" value="8"><label for="voxel8" style="background-position: -700% -0%"></label>
-      </div>
-      <div class="tiles">
-        <input type="radio" class="add" name="voxel" id="voxel9"  value="9" ><label for="voxel9"  style="background-position:  -800% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel10" value="10"><label for="voxel10" style="background-position:  -900% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel11" value="11"><label for="voxel11" style="background-position: -1000% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel12" value="12"><label for="voxel12" style="background-position: -1100% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel13" value="13"><label for="voxel13" style="background-position: -1200% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel14" value="14"><label for="voxel14" style="background-position: -1300% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel15" value="15"><label for="voxel15" style="background-position: -1400% -0%"></label>
-        <input type="radio" class="add" name="voxel" id="voxel16" value="16"><label for="voxel16" style="background-position: -1500% -0%"></label>
-      </div>
-    </div>
-    <div id="rightui">
-      <input type="button" class="button" id="modeButton" value="Editing Mode">
-      <input type="button" class="button" id="viewButton" value="Relativity OFF">
-      <input type="range" id="speedslider" min="0.01" max="0.99" value="0.7" step="0.01">
-      <h3 id="speedvalue">Max Speed: 0.70c</h3>
-      <input type="button" class="button" id="downloadButton" value="Save Your World">
-      <input type="button" class="button" id="loadfileButton" value="Load Saved World">
-
-      <div id="worldnameform">
-        <input type="text" id="worldnameinput" value="" placeholder="Set the name of world to save">
-        <input type="button" id="worldnamesubmit" value="Submit">
-      </div>
-
-    </div>
-    
-    <div id="bottomui">
-      <h3 style="color: white">Object to Edit</h3>
-      <input type="button" class="objectcreate" id="objectcreate" value="+">
-      <input type="button" class="objectdelete" id="objectdelete" value="-">
-    </div>
-    <div id="clocks">
-      <h3>Clocks</h3>
-      <input type="button" id="selectclock" value = "Select Clock To Delete">
-      <input type="button" id="deleteclock" value = "Delete Selected Clocks" style="display: none;">
-      <p class="time" id="mytime">My Time: </p>
-    </div>
-    <div id="clockform">
-      <input type="text" id="nameinput" value="" placeholder="Set the name of clock">
-      <input type="button" id="namesubmit" value="Submit">
-      <input type="hidden" id="intersection" value="">
-    </div>
-    <div id="objui">
-      <div>
-        vx: <input type="text" id="objvx" value="0.0">c
-      </div>
-      <div>
-        vy: <input type="text" id="objvy" value="0.0">c
-      </div>
-      <div>
-        vz: <input type="text" id="objvz" value="0.0">c
-      </div>
-      <input type="button" id="objsubmit" value="Set Object Velocity">
-    </div>
-		<script type="module" src="js/main.js"></script>
-
-	</body>
-</html>
