+++ package.json
@@ -1,0 +1,44 @@
+{
+  "name": "relativity-simulator",
+  "version": "1.0.0",
+  "description": "Relativity Simulator Source Code. Used Three.js for rendering tool. WEBSITE: https://relativity-simulator.netlify.app/",
+  "main": "index.js",
+  "scripts": {
+    "build": "webpack --config ./configs/webpack.config.prod.js",
+    "dev": "webpack serve --config ./configs/webpack.config.dev.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/seunghyun517/relativity-simulator.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/seunghyun517/relativity-simulator/issues"
+  },
+  "homepage": "https://github.com/seunghyun517/relativity-simulator#readme",
+  "dependencies": {
+    "dat.gui": "^0.7.9",
+    "three": "^0.139.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.17.8",
+    "@babel/preset-env": "^7.16.11",
+    "babel-loader": "^8.2.4",
+    "clean-webpack-plugin": "^4.0.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "css-loader": "^6.7.1",
+    "file-loader": "^6.2.0",
+    "html-loader": "^3.1.0",
+    "html-webpack-plugin": "^5.5.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "portfinder-sync": "^0.0.2",
+    "raw-loader": "^4.0.2",
+    "style-loader": "^3.3.1",
+    "webpack": "^5.70.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.7.4",
+    "webpack-merge": "^5.8.0"
+  }
+}
