+++ src/css/main.css
@@ -25,7 +25,7 @@ html, body {
   display: none;
 }
 #ui input[type=radio].add + label {
-  background-image: url("texture.png");
+  background-image: url("../../assets/img/texture.png");
   background-size: 1600% 400%;
   image-rendering: pixelated;
   width: 64px;
