+++ src/html/index.html
@@ -1,17 +1,18 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-eqUIv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Relativity Simulator</title>
-    <link rel="stylesheet" href="about.css">
+    <link rel="stylesheet" href="../css/about.css">
 
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=G-PF03MR37HK"></script>
     <script>
         window.dataLayer = window.dataLayer || [];
-        function gtag(){dataLayer.push(arguments);}
+        function gtag() { dataLayer.push(arguments); }
         gtag('js', new Date());
 
         gtag('config', 'G-PF03MR37HK');
@@ -20,23 +21,23 @@
 
 <body>
     <div>
-    <h1>Welcome to Relativity Simulator!</h1>
-    <h3>Created by Seunghyun Yeo</h3>
+        <h1>Welcome to Relativity Simulator!</h1>
+        <h3>Created by Seunghyun Yeo</h3>
 
-    <p>Here, you can have fantastic experiences in relativistic world. 
-        You can make a buildings just like Minecraft, create various objects moving in different velocity,
-    place clocks that really works, and watch how the special relativistic world look like. Please Enjoy!
-    </p>
+        <p>Here, you can have fantastic experiences in relativistic world.
+            You can make a buildings just like Minecraft, create various objects moving in different velocity,
+            place clocks that really works, and watch how the special relativistic world look like. Please Enjoy!
+        </p>
 
-    <a href="main.html">Start Relativity Simulator</a>
-    <br>
-    <a href="about.html">Read Instructions</a>
+        <a href="./main.html">Start Relativity Simulator</a>
+        <br>
+        <a href="./about.html">Read Instructions</a>
 
-    <br>
-    <br>
-    <img src="img/thumbnail.png">
+        <br>
+        <br>
+        <img src="../../assets/img/thumbnail.png">
 
-    
+
     </div>
     <footer>
         <p>Copyright 2022. Seunghyun Yeo All rights reserved.</p>
